attackRatio 적용, hp와 atk를 int에서 float로 바꾼다. (2)

+++ VampireSurvivors/app/src/main/java/kr/ac/tukorea/www/smgp2018184001/vampiresurvivors/game/characters/Character.java
@@ -68,7 +68,7 @@ public class Character extends Object implements ICollidable {
         reconstructColliderRect();
     }
 
-    public void getDamage(int damage) {
+    public void getDamage(float damage) {
         if (!isInvincible()) {
             aSprite.setIsInvincible(true);
             curHp -= damage;
